+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Updated tools:
 Misc:
 
 - Loosen unsatisfied constraints for npm packages [#2171](https://github.com/sider/runners/pull/2171)
-- Improve analysis finish message [#2187](https://github.com/sider/runners/pull/2187)
+- Improve analysis finish message [#2187](https://github.com/sider/runners/pull/2187) [#2232](https://github.com/sider/runners/pull/2232)
 - **ESLint** Provide new recommended configuration [#2165](https://github.com/sider/runners/pull/2165)
 - **ESLint** Use `--no-error-on-unmatched-pattern` if possible [#2199](https://github.com/sider/runners/pull/2199)
 - Simplify `sider.yml` schema [#2192](https://github.com/sider/runners/pull/2192)